Split take_order NB out of orderbook
